backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pandas>=2.2.0
qdrant-client>=1.17.0
requests>=2.32.0
python-dotenv>=1.0.0
